Add a hash to check completeness of lists
When the connection is cut during body transfer of a request to send lists
empty files may be write by send.php. To avoid that, the javascript now
sends a hash of the data as a GET parameter for send.php, thus corrupted
data is detected and asnwered by a 400 error.